Add root dropping
sshg-parser and sshg-blocker are sandboxed on FreeBSD and OpenBSD using
Capsicum and pledge(), respectively. However, on other systems SSHGuard
just runs with permissions of the user that started it, usually root.

Now SSHGuard also drops to the permissions of the user given by
SSHGUARD_USER or the USER variable in sshguard.conf.