Fix quoted string parser
Require matching pairs of quotes to parse a quoted string. This prevents
an unterminated quoted string (e.g. "foo) from keeping the parser in the
quoted_string lex state and triggering the default "echo characters to
standard output" rules.

This fixes a bug where unterminated strings as a result of having an CLF
request longer than MAX_BUF were echoed to the output of sshg-parser,
causing sshg-blocker to exit due to invalid inputs.