sshguard.in: Fix race condition in exit trap
When the sshguard script gets an SIGNAL to exit the clean_and_exit
function is called which calls exit which in turn calls exit 0. This can
in some cases call again clean_and_exit and causes a non-zero exit code.

The sh -x output looks like this in this cases:
...
+ /usr/libexec/sshg-fw-nft-sets
+ clean_and_exit
+ '[' -n  ]
+ exit
+ kill 0
+ clean_and_exit
+ '[' -n  ]
+ exit

~# echo $?
143

Disable the INT/TERM trap in clean_and_exit before exiting to allow a
clean exit.

Signed-off-by: Gregor Herburger <gregor.herburger@ew.tq-group.com>